fix(client, use): format repository name for valid environment variable name
When using the `trdl use group-name-with-dashes 1 stable` command, an invalid script was generated due to non-formated repository name. Now, the repository name is formatted to contain only letters, numbers, and underscores, ensuring it is a valid environment variable name.

All characters that are not letters, numbers, or underscores are replaced with underscores. The repository name is also converted to uppercase.

Signed-off-by: Aleksei Igrychev <aleksei.igrychev@palark.com>